requirements.txt
streamlit
opencv-python
numpy
torch
ultralytics
Pillow
pyyaml
